Refactor Page components
